# doc/api.yaml
openapi: 3.0.0
paths:
...
  '/favs/artist/{id}':
    {
      post:
        {
          operationId: FavsController_addArtistToFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '201': { description: 'Successfully added to favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '422': { description: 'Artist does not exist' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: FavsController_deleteArtistFromFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '204': { description: 'Successfully deleted from favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Artist is not favourite' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
  '/favs/album/{id}':
    {
      post:
        {
          operationId: FavsController_addAlbumToFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '201': { description: 'Successfully added to favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '422': { description: 'Album does not exist' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: FavsController_deleteAlbumFromFavs,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/FindFavsDto' } },
                },
            },
          responses:
            {
              '204': { description: 'Successfully deleted from favourites' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album is not favourite' },
            },
          tags: [favourites],
          security: [{ bearer: [] }],
        },
    }
  /album:
    {
      get:
        {
          operationId: AlbumsController_getAllEntities,
          parameters: [],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
      post:
        {
          operationId: AlbumsController_createAlbum,
          parameters: [],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/AlbumDto' } },
                },
            },
          responses:
            {
              '201':
                {
                  description: 'Successfully created',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '400': { description: 'Invalid dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
    }
  /album/search:
    {
      get:
        {
          operationId: AlbumsController_getAlbumsByName,
          parameters:
            [
              {
                name: name,
                required: true,
                in: query,
                description: 'Name of the album',
                schema: { type: string },
              },
            ],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
    }
  '/album/{id}':
    {
      get:
        {
          operationId: AlbumsController_getAlbumById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '404': { description: 'Album not found' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: AlbumsController_updateAlbum,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/AlbumDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Album' },
                            },
                        },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album not found' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: AlbumsController_deleteAlbum,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Album not found' },
            },
          tags: [albums],
          security: [{ bearer: [] }],
        },
    }
  /artist:
    {
      get:
        {
          operationId: ArtistsController_getAllArtists,
          parameters: [],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Artist' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
      post:
        {
          operationId: ArtistsController_createArtist,
          parameters: [],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/ArtistDto' } },
                },
            },
          responses:
            {
              '201':
                {
                  description: 'Successfully created',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Artist' } },
                    },
                },
              '400': { description: 'Invalid dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
    }
  '/artist/{id}':
    {
      get:
        {
          operationId: ArtistsController_getArtistById,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Artist' } },
                    },
                },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '404': { description: 'Artist not found' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
      put:
        {
          operationId: ArtistsController_updateArtist,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/ArtistDto' } },
                },
            },
          responses:
            {
              '200':
                {
                  description: 'Successfully updated',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Artist' } },
                    },
                },
              '400': { description: 'Invalid id or dto' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Artist not found' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
      delete:
        {
          operationId: ArtistsController_deleteArtist,
          parameters:
            [{ name: id, required: true, in: path, schema: { type: string } }],
          responses:
            {
              '204': { description: 'Successfully deleted' },
              '400': { description: 'Invalid id' },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
              '404': { description: 'Artist not found' },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
    }
  /artist/search:
    {
      get:
        {
          operationId: ArtistsController_getArtistsByName,
          parameters:
            [
              {
                name: name,
                required: true,
                in: query,
                description: 'Name of the artist',
                schema: { type: string },
              },
            ],
          responses:
            {
              '200':
                {
                  description: 'Successful operation',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: array,
                              items: { $ref: '#/components/schemas/Artist' },
                            },
                        },
                    },
                },
              '401': { description: Unauthorized },
            },
          tags: [artists],
          security: [{ bearer: [] }],
        },
    }
...
info:
  title: 'Home Library'
  description: 'Home music library service'
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes: { bearer: { scheme: bearer, bearerFormat: JWT, type: http } }
  schemas:
...
      FindFavsDto:
        {
          type: object,
          properties: { userId: { type: string } },
          required: [userId],
        },
...
      Album:
        {
          type: object,
          properties:
            {
              id: { type: string },
              name: { type: string },
              year: { type: number },
              artistId: { type: string },
            },
          required: [id, name, year, artistId],
        },
      AlbumDto:
        {
          type: object,
          properties:
            {
              name: { type: string },
              year: { type: number },
              artistId: { type: string },
            },
          required: [name, year],
        },
      Artist:
        {
          type: object,
          properties:
            {
              id: { type: string },
              name: { type: string },
              grammy: { type: boolean },
            },
          required: [id, name, grammy],
        },
      ArtistDto:
        {
          type: object,
          properties: { name: { type: string }, grammy: { type: boolean } },
          required: [name, grammy],
        },
...