# src/albums/dto/album.dto.ts
import { IsNumber, IsOptional, IsString, IsNotEmpty } from 'class-validator';
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class AlbumDto {
  @ApiProperty({ type: String })
  @IsString()
  @IsNotEmpty()
  name: string;

  @ApiProperty({ type: Number })
  @IsNumber()
  @IsNotEmpty()
  year: number;

  @ApiPropertyOptional({ type: String })
  @IsString()
  @IsOptional()
  artistId: string | null;
}